+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - JUnit extension to collect TableTest report data during test runs (tabletest-reporter-junit)
 - Multiple roles supported per cell in published YAML
 - Roles added to signal if a row passed or failed
+- CamelCase and snake_case aware slugified YAML file name generation
 ### Changed
 - YAML files prefixed with `TABLETEST-` to avoid conflicts with other YAML files
 - Output file name for TableTest methods either explicit `@DisplayName` or method name (without parameters)